requirements.txt
absl-py==1.0.0
aiohttp==3.8.0
aiosignal==1.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
bayesian-optimization==1.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
cma==3.2.2
configparser==5.2.0
ConfigSpace==0.4.20
cycler==0.11.0
Cython==0.29.24
dask==2021.11.1
dask-glm==0.2.0
dask-ml==2021.11.30
datasets==1.15.1
Deprecated==1.2.13
dill==0.3.4
distributed==2021.11.1
docker-pycreds==0.4.0
emoji==1.6.1
fasttext==0.9.2
filelock==3.3.2
flatbuffers==2.0
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
HeapDict==1.0.1
huggingface-hub==0.1.2
hyperopt==0.2.7
idna==3.3
imageai==2.1.6
importlib-metadata==4.8.2
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.2
libclang==12.0.0
llvmlite==0.37.0
locket==0.2.1
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.2
msgpack==1.0.3
multidict==5.2.0
multipledispatch==0.6.0
multiprocess==0.70.12.2
networkx==2.6.3
nevergrad==0.5.0
nlpaug==1.1.10
nltk==3.6.5
numba==0.54.1
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
Owlready2==0.35
packaging==21.3
pandas==1.3.4
partd==1.2.0
pathtools==0.1.2
patsy==0.5.2
pickle5==0.0.12
Pillow==7.0.0
pipelinehelper==0.7.8
pkg_resources==0.0.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
py4j==0.10.9.3
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pydot==1.4.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
ray==1.9.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.4.1
seaborn==0.11.2
sentry-sdk==1.5.2
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
sortedcontainers==2.4.0
stanza==1.3.0
statsmodels==0.13.2
subprocess32==3.5.4
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toolz==0.11.2
torch==1.10.0
tornado==6.1
tqdm==4.62.3
transformers==4.12.3
typeguard==2.13.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.9
Werkzeug==2.0.2
wordcloud==1.8.1
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
zict==2.0.0
zipp==3.6.0
